% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_create.R
\name{cluster_exists}
\alias{cluster_exists}
\title{Checks if a cluster exists.}
\usage{
cluster_exists(clustername)
}
\arguments{
\item{clustername}{Name of the cluster}
}
\description{
Checks if a cluster exists.
}
