% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_connect.R
\name{connect}
\alias{connect}
\title{connect}
\usage{
connect(settingspath)
}
\arguments{
\item{settingspath}{Location where the settings.json file is stored.}
}
\description{
Connects to the Ravana database.
Connection settings are retrieved using the load_settings function.
Settings are stored in the settings.json file.
}
