% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_create.R
\name{init_cluster}
\alias{init_cluster}
\title{Initializes a Ravana cluster and creates an entry in the database.
If a cluster with the same name already exists, this process will load related functions and variables from the database.}
\usage{
init_cluster(clustername, settingspath = ".")
}
\arguments{
\item{clustername}{Name of the cluster}

\item{settingspath}{Location where the settings.json file is stored. If no path is provided it will look into the working directory (getwd()).}
}
\description{
Initializes a Ravana cluster and creates an entry in the database.
If a cluster with the same name already exists, this process will load related functions and variables from the database.
}
