% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_connect.R
\name{load_settings}
\alias{load_settings}
\title{load_settings}
\usage{
load_settings(settingspath)
}
\arguments{
\item{settingspath}{Location where the settings.json file is stored.}
}
\description{
Loads PostgreSQL database connection settings from the settings.json file.
If the file not found, an error will be raised.
}
