% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana.R
\name{Map}
\alias{Map}
\title{Map a Registered R Function}
\usage{
Map(Function, Data, BatchSize = 1)
}
\arguments{
\item{Name}{of the registered R Function}

\item{List}{containing the data}

\item{Batch}{size}
}
\value{
\link{Void}
}
\description{
Map a Registered R Function
}
