% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_mapreduce.R
\name{ravana_map}
\alias{ravana_map}
\title{ravana_map}
\usage{
ravana_map(rfunction, datatomap)
}
\arguments{
\item{rfunction}{R function}

\item{mappeddata}{}
}
\value{
Returns the taskid.
}
\description{
Maps a task expressed as an R function to each element of list.
}
