% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_mapreduce.R
\name{ravana_reduce}
\alias{ravana_reduce}
\title{Reduces a task which has been already Mapped.}
\usage{
ravana_reduce(taskid)
}
\arguments{
\item{taskid}{}

\item{rfunction}{R function}
}
\value{
Returns the results in a list.
}
\description{
Reduces a task which has been already Mapped.
}
