% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_create.R
\name{set_worker}
\alias{set_worker}
\title{set_worker}
\usage{
set_worker()
}
\description{
Sets the current computer as a worker node.
Workers receive tasks, completes them using suggested functions and return the result back to the database.
}
