% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_mapreduce.R
\name{share_function}
\alias{share_function}
\title{Shares an R function in the cluster}
\usage{
share_function(rfunction)
}
\arguments{
\item{function}{R function}
}
\description{
Shares an R function in the cluster
}
