% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_mapreduce.R
\name{share_object}
\alias{share_object}
\title{Shares an R object (vector, list, matrix, data frame etc.) in the cluster}
\usage{
share_object(robject)
}
\arguments{
\item{robject}{R object}
}
\description{
Shares an R object (vector, list, matrix, data frame etc.) in the cluster
}
