% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravana_create.R
\name{resume_cluster}
\alias{resume_cluster}
\title{resume_cluster}
\usage{
resume_cluster(clustername)
}
\arguments{
\item{clustername}{Name of the cluster}
}
\description{
Resumes an existing cluster by loading details from the database.
}
